ray_cluster/requirements.cpu.txt
torch --index-url https://download.pytorch.org/whl/cpu
ray
transformers
psutil
PyYAML 